+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Basic BWT for sub-sequence count and offset for sequence alignment. Only supports exact matches for now.
+
 
 ## [0.30.0] - 2023-12-18
 Oops, we weren't keeping a changelog before this tag!